chore(defuse-borsh-utils): schema feature
