+++ contracts/defuse/src/contract/versioned/mod.rs
@@ -5,7 +5,7 @@ use std::{
     io::{self, Read},
 };
 
-use defuse_borsh_utils::adapters::{BorshDeserializeAs, BorshSerializeAs};
+use defuse_borsh_utils::{BorshDeserializeAs, BorshSerializeAs};
 use defuse_near_utils::PanicOnClone;
 use near_sdk::{
     borsh::{BorshDeserialize, BorshSerialize},
