+++ contracts/wallet/src/signature/mod.rs
@@ -10,7 +10,7 @@ pub mod webauthn;
 
 use std::time::Duration;
 
-use defuse_borsh_utils::adapters::{As, DurationSeconds as BorshDurationSeconds};
+use defuse_borsh_utils::{As, DurationSeconds as BorshDurationSeconds};
 use defuse_time::{Timestamp, borsh::TimestampSeconds};
 use near_sdk::{AccountId, CryptoHash, env, near, serde_with::DurationSeconds};
